Restructure

+++ src/kernel/arch/x86_64/context.h
@@ -1,7 +1,7 @@
 #pragma once
 
-#include <lithe/base/defs.h>
-#include <lithe/base/attributes.h>
+#include <utils/base/attributes.h>
+#include <utils/base/defs.h>
 
 typedef struct {
 	uint64_t r15, r14, r13, r12, r11, r10, r9, r8, rbp, rdi, rsi, rdx, rcx,
@@ -9,7 +9,7 @@ typedef struct {
 } PACKED regs_t;
 
 typedef struct {
-	uintptr_t kernel_stack;
-	uintptr_t user_stack;
+	uint64_t kernel_stack;
+	uint64_t user_stack;
 	regs_t regs;
 } context_t;
